enhancement: refactoring project structure and pep8 format
